+++ changelog.md
@@ -171,7 +171,7 @@ its results in a DBF file that is read by the MLogit.R script. "RJDBC" is indeed
     - (Partially) language (Locale) change without restarting Nodus.
     - No more exception thrown when statistics are gathered for a non existing group in StatPieDlg.
     - Limit max heap size to 1.4Go if run on a 32bit JVM
-    - Limit the number of threads for an assignment to the number of physical cores (not including 'hyperthreaded' cores.
+    - Warn the user if the number of threads set for an assignment is larger than the number of physical cores.
 
 - Breaking changes:
     - Simplify the API for modal split methods. This breaks existing plugin's. If an incompatible plugin is found, an error